Revert to using string for ST_Geometry values.
While https://github.com/azavea/net-dao-odp/pull/5 improved the way that
ST_Geometry column values were detected, it failed to set the value
of the parameter correctly. This resulted in ST_Geometry columns
being populated with null values.

Here we leave the ST_Geometry value check unchanged, but revert to
setting the value to the string representation of the ST_Geometry.
A different problem with the ULRS config files was also causing
INSERT issues, and it turns out that this did not need to be updated.

Refs https://github.com/azavea/oit-ulrs/issues/180.